Do not pass the next query param to Django
The Django login view does special handling of the ?next param, but that
doesn't make sense to do in this setup since the url will only exist on
the frontend. So we make sure to not pass query parameters through the
PassthroughForm and handle the next logic only on the frontend in this
case.